+++ task3/public/js/cat.js
