+++ task3/public/js/hasher.js
