+++ task3/public/js/joke.js
