+++ task3/public/js/messageRedirect.js
