+++ task3/public/js/pathCreator.js
@@ -1,11 +1,11 @@
 const path = require("path");
 
 
-const staticPath = path.resolve(__dirname, "../static");
+const staticPath = path.resolve(__dirname, "../../static");
 
 
 const createViewPath = (page) =>
-  path.resolve(__dirname, "../views", `${page}.ejs`);
+  path.resolve(__dirname, "../../src/views", `${page}.ejs`);
 
 module.exports = {
   path,
