+++ task3/public/js/static.js
