+++ task3/src/views/partials/includeCss.ejs
@@ -1,0 +1,2 @@
+<link rel="stylesheet" href="/css/fileDirectory.css">
+<link rel="stylesheet" href="/css/header.css">
