+++ task3/src/views/partials/viewJokes.ejs
