+++ task3/src/views/singUp.ejs
