--- task3/test.sql
@@ -1,1 +1,0 @@
-DELETE FROM users
