--- task3/views/partials/includeCss.ejs
@@ -1,2 +1,0 @@
-<link rel="stylesheet" href="/fileDirectory.css">
-<link rel="stylesheet" href="/header.css">
